add: enable global exception  handler add: enable api versioning refactor: move mapping logic to mapper class

+++ studentsapi/src/main/java/com/scotiabank/studentsapi/model/dto/StudentResponse.java
@@ -14,5 +14,5 @@ public class StudentResponse {
     private String name;
     private String surname;
     private String status;
-    private Integer age;
+    private Short age;
 }
